stop let-generalization for now...
I can't figure out an algorithm for monomorphization. I have figured out the general rules and the input and output states, yet the algorithm seems too hard. I kinda want to use my language tbh. Let-generalization is a nice bragging right (expressibility of dynamic typed languages with type system of Haskell and performance of C), but it would take a lot of time...
... also I can easily add it if the time comes...