Another typeclassing step. RemoveUnused was removed.
Was it a good idea? We'll see.
I'll have to think more about the architecture, but currently I think we've simplified stuff a bit.
We don't need the 'cuckedUnions' field anymore, which was already a smell.

Right now, we replace every IncompleteEnv with EnvID and match on UnionID. This is because we might encounter an Env in a type, which was not yet type mapped. Also, we might need to eliminate environment members, because they are not used (previously delegated to RemoveUnused).

Also, I removed 'RemoveUnused' file, because it was getting (or it would get with full introduction of typeclasses) too complex and
was basically doing its own monomorphising (for example, we can't tell which instance function gets used until after we monomorphize and this might require deep environment updates). Regarding 'cuckedUnions', removing the module seemed to simplify this aspect as well.

Currently, source files are not very clean with commented out code. I'll clean it up later.

But I think I'm up to speed on the project again.